Move config.py to __init__.py
