+++ esque/clients/producer.py
@@ -8,7 +8,7 @@ import confluent_kafka
 import pendulum
 from confluent_kafka.avro import AvroProducer
 
-from esque.config.config import Config
+from esque.config import Config
 from esque.errors import raise_for_kafka_error
 from esque.helpers import delivery_callback, delta_t
 from esque.messages.avromessage import AvroFileReader
