+++ esque/cluster.py
@@ -3,7 +3,7 @@ import operator
 import pykafka
 from confluent_kafka.admin import AdminClient, ConfigResource
 
-from esque.config.config import Config
+from esque.config import Config
 from esque.controller.topic_controller import TopicController
 from esque.helpers import ensure_kafka_futures_done, unpack_confluent_config
